requirements.txt
requests
pandas
python-dotenv
sentence-transformers
faiss-cpu
PyMuPDF
streamlit
